move examples into accordion
